feat(xml, yaml): add custom XmlBeanPropertyWriter and YAMLParser for enhanced serialization

+++ src/main/java/com/norconex/commons/lang/time/DurationParser.java
@@ -204,7 +204,7 @@ public class DurationParser {
                         "Invalid duration value: \"%s\".", duration, numGroup);
                 return defaultValue;
             }
-            var val = NumberUtils.toFloat(num, -1);
+            var val = NumberUtils.toDouble(num, -1);
             if (val == -1) {
                 parseError(throwException,
                         "Invalid duration value: \"%s\".", duration, numGroup);
